Don't let updates clobber scrolling
When we get an update, we set ViewPosition. This will in theory be used
to position the next rendered frame via CheckTargetUpdate(). However it
can be some time before CheckViewPosition() actually gets as far as
scrolling to the saved position, because of recalculations.

So we need to check whether there has been any scrolling since we saved
ViewPosition.

This adds an additional bit of state, savedTopOffset, which is the top
cursor position of the top page at the time of generating ViewPosition.
Probably it should be on one of the State/ViewPosition Anchor values.